reconfigure mina-indexer.rs and server.rs
