+++ src/bin/mina-indexer.rs
@@ -1,11 +1,15 @@
 use clap::{Parser, Subcommand};
 use mina_indexer::{
+    MAINNET_GENESIS_HASH, MAINNET_TRANSITION_FRONTIER_K, PRUNE_INTERVAL_DEFAULT, CANONICAL_UPDATE_THRESHOLD,
     client,
-    server::{self, create_dir_if_non_existent, handle_command_line_arguments, MinaIndexer},
-    store::IndexerStore,
+    server::{create_dir_if_non_existent, MinaIndexer, IndexerConfiguration},
+    store::IndexerStore, block::BlockHash, state::ledger,
 };
+use serde::Deserializer;
+use serde_derive::Deserialize;
+use tracing::{trace, instrument, info, error};
 use std::{path::PathBuf, sync::Arc};
-use tracing_subscriber::prelude::*;
+use tracing_subscriber::{prelude::*, filter::LevelFilter};
 
 #[derive(Parser, Debug)]
 #[command(name = "mina-indexer", author, version, about, long_about = Some("Mina Indexer\n\n\
@@ -40,7 +44,56 @@ enum ServerCommand {
         path: PathBuf,
     },
     /// Start the mina indexer by passing in arguments manually on the command line
-    Cli(server::ServerArgs),
+    Cli(ServerArgs),
+}
+
+#[derive(Parser, Debug, Clone, Deserialize)]
+#[command(author, version, about, long_about = None)]
+pub struct ServerArgs {
+    /// Path to the root ledger (if non-genesis, set --non-genesis-ledger and --root-hash)
+    #[arg(short, long)]
+    ledger: PathBuf,
+    /// Use a non-genesis ledger
+    #[arg(short, long, default_value_t = false)]
+    non_genesis_ledger: bool,
+    /// Hash of the base ledger
+    #[arg(
+        long,
+        default_value = MAINNET_GENESIS_HASH
+    )]
+    root_hash: String,
+    /// Path to startup blocks directory
+    #[arg(short, long, default_value = concat!(env!("HOME"), "/.mina-indexer/startup-blocks"))]
+    startup_dir: PathBuf,
+    /// Path to directory to watch for new blocks
+    #[arg(short, long, default_value = concat!(env!("HOME"), "/.mina-indexer/watch-blocks"))]
+    watch_dir: PathBuf,
+    /// Path to directory for rocksdb
+    #[arg(short, long, default_value = concat!(env!("HOME"), "/.mina-indexer/database"))]
+    pub database_dir: PathBuf,
+    /// Path to directory for logs
+    #[arg(long, default_value = concat!(env!("HOME"), "/.mina-indexer/logs"))]
+    pub log_dir: PathBuf,
+    /// Only store canonical blocks in the db
+    #[arg(short, long, default_value_t = false)]
+    keep_non_canonical_blocks: bool,
+    /// Max file log level
+    #[serde(deserialize_with = "level_filter_deserializer")]
+    #[arg(long, default_value_t = LevelFilter::DEBUG)]
+    pub log_level: LevelFilter,
+    /// Max stdout log level
+    #[serde(deserialize_with = "level_filter_deserializer")]
+    #[arg(long, default_value_t = LevelFilter::INFO)]
+    pub log_level_stdout: LevelFilter,
+    /// Interval for pruning the root branch
+    #[arg(short, long, default_value_t = PRUNE_INTERVAL_DEFAULT)]
+    prune_interval: u32,
+    /// Threshold for updating the canonical tip/ledger
+    #[arg(short, long, default_value_t = CANONICAL_UPDATE_THRESHOLD)]
+    canonical_update_threshold: u32,
+    /// Path to an indexer snapshot
+    #[arg(long)]
+    pub snapshot_path: Option<PathBuf>,
 }
 
 #[tokio::main]
@@ -97,4 +150,96 @@ pub async fn main() -> anyhow::Result<()> {
             Ok(())
         }
     }
+}
+
+#[instrument(skip_all)]
+pub async fn handle_command_line_arguments(
+    args: ServerArgs,
+) -> anyhow::Result<IndexerConfiguration> {
+    trace!("Parsing server args");
+
+    let ledger = args.ledger;
+    let non_genesis_ledger = args.non_genesis_ledger;
+    let root_hash = BlockHash(args.root_hash.to_string());
+    let startup_dir = args.startup_dir;
+    let watch_dir = args.watch_dir;
+    let keep_noncanonical_blocks = args.keep_non_canonical_blocks;
+    let prune_interval = args.prune_interval;
+    let canonical_update_threshold = args.canonical_update_threshold;
+
+    assert!(
+        ledger.is_file(),
+        "Ledger file does not exist at {}",
+        ledger.display()
+    );
+    assert!(
+        // bad things happen if this condition fails
+        canonical_update_threshold < MAINNET_TRANSITION_FRONTIER_K,
+        "canonical update threshold must be strictly less than the transition frontier length!"
+    );
+
+    create_dir_if_non_existent(watch_dir.to_str().unwrap()).await;
+
+    info!("Parsing ledger file at {}", ledger.display());
+
+    match ledger::genesis::parse_file(&ledger).await {
+        Err(err) => {
+            error!(
+                reason = "Unable to parse ledger",
+                error = err.to_string(),
+                path = &ledger.display().to_string()
+            );
+            std::process::exit(100)
+        }
+        Ok(ledger) => {
+            info!("Ledger parsed successfully!");
+
+            Ok(IndexerConfiguration {
+                ledger,
+                non_genesis_ledger,
+                root_hash,
+                startup_dir,
+                watch_dir,
+                keep_noncanonical_blocks,
+                prune_interval,
+                canonical_update_threshold,
+                from_snapshot: args.snapshot_path.is_some(),
+            })
+        }
+    }
+}
+
+pub fn level_filter_deserializer<'de, D>(deserializer: D) -> Result<LevelFilter, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct YAMLStringVisitor;
+
+    impl<'de> serde::de::Visitor<'de> for YAMLStringVisitor {
+        type Value = LevelFilter;
+
+        fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
+            formatter.write_str("a string containing yaml data")
+        }
+
+        fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+        where
+            E: serde::de::Error,
+        {
+            // unfortunately we lose some typed information
+            // from errors deserializing the json string
+            let level_filter_str: &str = serde_yaml::from_str(v).map_err(E::custom)?;
+            match level_filter_str {
+                "info" => Ok(LevelFilter::INFO),
+                "debug" => Ok(LevelFilter::DEBUG),
+                "error" => Ok(LevelFilter::ERROR),
+                "trace" => Ok(LevelFilter::TRACE),
+                "warn" => Ok(LevelFilter::TRACE),
+                "off" => Ok(LevelFilter::OFF),
+                other => Err(E::custom(format!("{} is not a valid level filter", other))),
+            }
+        }
+    }
+
+    deserializer.deserialize_any(YAMLStringVisitor)
 }
